requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# LangGraph + LangChain
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-ollama>=1.0.0
langchain-openai>=1.0.0

# ML
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.4.0

# Database
sqlalchemy>=2.0.0

# Frontend
streamlit>=1.35.0
pandas>=2.2.0
requests>=2.31.0

# Config
python-dotenv>=1.0.0
